+++ framework/py/flwr/supercore/interceptors/http/__init__.py
@@ -1,0 +1,20 @@
+# Copyright 2026 Flower Labs GmbH. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ==============================================================================
+"""Reusable protobuf-over-HTTP client interceptors."""
+
+from .runtime_version import RuntimeVersionHttpInterceptor
+from .superexec_auth import SuperExecAuthHttpInterceptor
+
+__all__ = ["RuntimeVersionHttpInterceptor", "SuperExecAuthHttpInterceptor"]
